+++ api/service/BoxService.cs
@@ -17,12 +17,12 @@ public class BoxService
     {
         return _boxRepository.GetBoxStock();
     }
-    public Box CreateBox(string size,int weight, int price, string material, string color, int quantity)
+    public Box CreateBox(string size,float weight, float price, string material, string color, int quantity)
     {
         return _boxRepository.CreateBox(size, weight, price, material, color, quantity);
     }
 
-    public Box UpdateBox(int id, string size, int weight, int price, string material, string color, int quantity)
+    public Box UpdateBox(int id, string size, float weight, float price, string material, string color, int quantity)
     {
         return _boxRepository.UpdateBox(id, size, weight, price, material, color, quantity);
     }
